code정리 + chat포트 하나만 열리게
