+++ server/.babelrc
@@ -1,7 +1,7 @@
 {
     "presets": [
         [
-            "@babel/preset-env",
+            "@babel/preset-env"
         ]
     ]
 }
