--- server/db/db2.js
@@ -1,9 +1,0 @@
-let mysql = require('mysql');
-let db2 = mysql.createConnection({
-  host: '3.35.211.102',
-  user: 'root',
-  password: '1122',
-  database: 'together'
-});
-
-module.exports=db2;
