+++ server/package.json
@@ -18,17 +18,17 @@
     "ejs": "^3.1.6",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
-    "http": "^0.0.1-security",
     "multer": "^1.4.2",
     "mysql": "^2.18.1",
+    "mysql2": "^2.3.0",
     "passport": "^0.4.1",
     "passport-kakao": "^1.0.1",
     "passport-naver": "^1.0.6",
     "request": "^2.88.2",
-    "websocket": "^1.0.34",
     "ws": "^7.4.6"
   },
   "devDependencies": {
+    "@babel/cli": "^7.14.8",
     "nodemon": "^2.0.7"
   }
 }
